requirements.txt
dvc
feast
matplotlib
mlflow
pandas
pyarrow
pymongo
python-dotenv
requests
scikit-learn
seaborn